fixed JWKS/export and validation issue
