+++ apps/authx_token_api/vitest.config.ts
@@ -1,4 +1,4 @@
-import { defineWorkersConfig } from "@cloudflare/vitest-pool-workers/config";
+/*import { defineWorkersConfig } from "@cloudflare/vitest-pool-workers/config";
 
 export default defineWorkersConfig({
   test: {
@@ -8,4 +8,11 @@ export default defineWorkersConfig({
       },
     },
   },
-});
+});*/
+
+import { defineConfig } from 'vitest/config'
+export default defineConfig({
+  test: {
+    exclude: [],
+  },
+})
