+++ apps/data_channel_gateway/tests/jwt.spec.ts
@@ -2,7 +2,7 @@
 
 import { describe, expect, it } from 'vitest';
 import { env } from 'cloudflare:test';
-import {importJWK, jwtVerify} from "jose"
+import {importJWK, jwtVerify, createLocalJWKSet} from "jose"
 describe("jwt integration tests", () => {
   it("can get the public key", async () =>{
     console.log(env)
@@ -52,17 +52,17 @@ describe("jwt integration tests", () => {
   it("can use jwks", async () => {
     const jwtDoId = env.JWT_TOKEN_DO.idFromName("default")
     const jwtStub = env.JWT_TOKEN_DO.get(jwtDoId)
-    const jwk = await jwtStub.getPublickKeyJWK()
+    const jwk = await jwtStub.getJWKS()
     console.log(jwk)
     expect(jwk).toBeDefined()
-    const publicKey = await importJWK(jwk)
     const jwtRequest = {
       entity: "testuser",
       claims: ["testclaim"],
     }
     const jwtToken = await jwtStub.signJWT(jwtRequest, 1000000000)
 
-  const { payload, protectedHeader } = await jwtVerify(jwtToken.token, publicKey, {
+    const jwkPub = await createLocalJWKSet(jwk);
+  const { payload, protectedHeader } = await jwtVerify(jwtToken.token, jwkPub, {
     issuer: 'catalyst:system:jwt:latest',
     audience: 'catalyst:system:datachannels',
   })
